second major redesign
Scott Harden:

pySquelch_graph.py - only responsible for reading the log file and
generating (new, fancier) graphs and HTML. Easy to swap out as it's
upgraded.  Change it / test it all you want.  It won't mess up your
logging.

pySquelch_log.py - only responsible for putting audio data into the log
file.  This DOES NOT LOOP and has to be called (by a cron job every
minute).  This way it completely closes when it's done, eliminating the
possibility of out-of-memory errors.  If it finds a new audio file, it
analyzes it, logs it, and launches the graphing script itself.  (you
don't need to add a cron job for the grapher)